LPS-51819 build services

+++ portal-service/src/com/liferay/portal/service/OrganizationLocalService.java
@@ -1170,7 +1170,7 @@ public interface OrganizationLocalService extends BaseLocalService,
 	attributes for the organization.
 	* @return the organization
 	* @deprecated As of 7.0.0, replaced by {@link #updateOrganization(long,
-	long, long, String, String, long, long, int, String, boolean,
+	long, long, String, String, long, long, long, String, boolean,
 	byte[], boolean, ServiceContext)}
 	*/
 	@java.lang.Deprecated
